@@ -1,8 +1,8 @@
 <Project>
 	<PropertyGroup>
 		<Version>1.0.0-pre.0.3</Version>
-		<Copyright>© Assimalign LLC 2021</Copyright>
-		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+		<Copyright>© Assimalign LLC 2022</Copyright>
+		<TargetFrameworks>net6.0</TargetFrameworks>
 		<LangVersion>10.0</LangVersion>
 		<Authors>Chase R. Crawford</Authors>
 		<PackageIconUrl>https://corporate.assets.assimalign.com/assets/logos/.assimalign.logo.png</PackageIconUrl>
