@@ -9,10 +9,10 @@ namespace Assimalign.ComponentModel.Mapping;
 
 using Assimalign.ComponentModel.Mapping.Abstractions;
 
-internal sealed class MapperProfileDefault<TSource, TTarget> : MapperProfile<TSource, TTarget>
+internal sealed class MapperProfileDefault<TContract, TBinding> : MapperProfile<TContract, TBinding>
 {
 
-    public override void Configure(IMapperProfileDescriptor<TSource, TTarget> descriptor)
+    public override void Configure(IMapperProfileDescriptor<TContract, TBinding> descriptor)
     {
         throw new NotImplementedException();
     }
